@@ -7,11 +7,11 @@
         "php": ">=8.2",
         "ext-ctype": "*",
         "ext-iconv": "*",
-        "doctrine/dbal": "^3",
-        "doctrine/doctrine-bundle": "^2.13",
+        "doctrine/dbal": "^3.9.4",
+        "doctrine/doctrine-bundle": "^2.13.2",
         "doctrine/doctrine-migrations-bundle": "^3.4",
-        "doctrine/orm": "^3.3",
-        "phpdocumentor/reflection-docblock": "^5.6",
+        "doctrine/orm": "^3.3.1",
+        "phpdocumentor/reflection-docblock": "^5.6.1",
         "phpstan/phpdoc-parser": "^2.0",
         "symfony/asset": "7.2.*",
         "symfony/asset-mapper": "7.2.*",
@@ -19,14 +19,14 @@
         "symfony/doctrine-messenger": "7.2.*",
         "symfony/dotenv": "7.2.*",
         "symfony/expression-language": "7.2.*",
-        "symfony/flex": "^2",
+        "symfony/flex": "^2.4.7",
         "symfony/form": "7.2.*",
         "symfony/framework-bundle": "7.2.*",
         "symfony/http-client": "7.2.*",
         "symfony/intl": "7.2.*",
         "symfony/mailer": "7.2.*",
         "symfony/mime": "7.2.*",
-        "symfony/monolog-bundle": "^3.0",
+        "symfony/monolog-bundle": "^3.10",
         "symfony/notifier": "7.2.*",
         "symfony/process": "7.2.*",
         "symfony/property-access": "7.2.*",
@@ -34,16 +34,16 @@
         "symfony/runtime": "7.2.*",
         "symfony/security-bundle": "7.2.*",
         "symfony/serializer": "7.2.*",
-        "symfony/stimulus-bundle": "^2.22",
+        "symfony/stimulus-bundle": "^2.22.1",
         "symfony/string": "7.2.*",
         "symfony/translation": "7.2.*",
         "symfony/twig-bundle": "7.2.*",
-        "symfony/ux-turbo": "^2.22",
+        "symfony/ux-turbo": "^2.22.1",
         "symfony/validator": "7.2.*",
         "symfony/web-link": "7.2.*",
         "symfony/yaml": "7.2.*",
-        "twig/extra-bundle": "^2.12|^3.0",
-        "twig/twig": "^2.12|^3.0"
+        "twig/extra-bundle": "^2.12|^3.18",
+        "twig/twig": "^2.12|^3.18"
     },
     "config": {
         "allow-plugins": {
@@ -85,6 +85,18 @@
         ],
         "post-update-cmd": [
             "@auto-scripts"
+        ],
+        "lint": "pint",
+        "refactor": "rector",
+        "test:unit": "php bin/phpunit",
+        "test:lint": "pint --test",
+        "test:types": "phpstan analyse --ansi",
+        "test:refactor": "rector --dry-run",
+        "test": [
+            "@test:unit",
+            "@test:lint",
+            "@test:types",
+            "@test:refactor"
         ]
     },
     "conflict": {
@@ -97,11 +109,14 @@
         }
     },
     "require-dev": {
-        "phpunit/phpunit": "^9.5",
+        "laravel/pint": "^1.20",
+        "phpstan/phpstan": "^2.1",
+        "phpunit/phpunit": "^9.6.22",
+        "rector/rector": "^2.0",
         "symfony/browser-kit": "7.2.*",
         "symfony/css-selector": "7.2.*",
         "symfony/debug-bundle": "7.2.*",
-        "symfony/maker-bundle": "^1.0",
+        "symfony/maker-bundle": "^1.62.1",
         "symfony/phpunit-bridge": "^7.2",
         "symfony/stopwatch": "7.2.*",
         "symfony/web-profiler-bundle": "7.2.*"
